Add persistent volume claims for storage in EKS and GKE deployments (#827)
* Add persistent volume claims for storage and logs in EKS and GKE deployments

- Updated kustomization files to include mpm-storage-pvc.yaml.
- Added volume mounts and volume definitions for storage and logs in mpm-backend, mpm-scheduler, and GKE deployments.

* fix yamllint errors

* persist storage directory with volume directive on dockerfiles.

Persist /var/www/html/storage folder with the VOLUME directive on backend docker images
The make is visible to anyone running these images that the storage folder is expected to be persisted

* update dev docker compose

* mount shared storage on the host system

* fix lint warnings

* move static files out of storage directory

* update resolution of assets

* removing realpath fixed startup error on laravel when storage was empty

* mount storage volume on docker servies

* Fix linting errors

* cleanup storage folder

* set storage folder permissions, create folders on startup

* code refactorings

Move centralized logic for created sub-folders in storage folder needed for portions of the app to the areas that need them i.e check for folder existence before use.
Include volume directive to Dockerfiles and specify volumes on prod docker compse and dockerhub files.

* add template files